+++ src/errors.js
@@ -17,7 +17,7 @@ export class StorageError extends Error {
   /**
    * Error with status code
    * @param {string} message Error message
-   * @param {string} code Error code
+   * @param {string} [code] Error code
    */
   constructor(message, code) {
     super(message);